requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-storage>=2.1.0
google-cloud-documentai>=1.2.0
google-cloud-bigquery>=2.32.0
pandas>=1.3.0
pyarrow>=6.0.0